# palabatu-be/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "internal_problems.CreateProblemRequest": {
                "properties": {
                    "boulder_id": {
                        "type": "string"
                    },
                    "descent": {
                        "type": "string"
                    },
                    "discovered_by": {
                        "type": "string"
                    },
                    "first_ascensionist": {
                        "type": "string"
                    },
                    "grade": {
                        "type": "string"
                    },
                    "height_m": {
                        "type": "number"
                    },
                    "landing_hazards": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_problems.ProblemListItem": {
                "properties": {
                    "boulder_id": {
                        "type": "string"
                    },
                    "boulder_name": {
                        "type": "string"
                    },
                    "crag_id": {
                        "type": "string"
                    },
                    "crag_name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "creator_name": {
                        "type": "string"
                    },
                    "creator_slug": {
                        "type": "string"
                    },
                    "descent": {
                        "type": "string"
                    },
                    "discovered_by": {
                        "type": "string"
                    },
                    "first_ascensionist": {
                        "type": "string"
                    },
                    "grade": {
                        "type": "string"
                    },
                    "height_m": {
                        "type": "number"
                    },
                    "id": {
                        "type": "string"
                    },
                    "landing_hazards": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "send_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "palabatu-be_internal_apitypes.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "palabatu-be_internal_apitypes.SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and the JWT, e.g. \"Bearer eyJhbGc...\".",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Backend API for Palabatu, a community app for Indonesian bouldering enthusiasts (spot map, climber profiles, route/problem listings, comments, reactions, reports).",
        "title": "Palabatu API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/problems": {
            "get": {
                "description": "Optional crag_id/boulder_id query params filter to one crag or boulder.",
                "parameters": [
                    {
                        "description": "Filter to one crag",
                        "in": "query",
                        "name": "crag_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter to one boulder",
                        "in": "query",
                        "name": "boulder_id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/internal_problems.ProblemListItem"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List all problems",
                "tags": [
                    "problems"
                ]
            },
            "post": {
                "description": "Any authenticated user may create a problem; no role gate. boulder_id is required -- crag_id is derived from the boulder, not supplied directly.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_problems.CreateProblemRequest",
                                        "summary": "body",
                                        "description": "New problem"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "New problem",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_problems.ProblemSummary"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a problem",
                "tags": [
                    "problems"
                ]
            }
        },
        "/api/problems/{id}": {
            "delete": {
                "description": "Allowed for admins (Council/Associate title) on any problem, or the problem's own creator. Does not touch the boulder's shared photos -- only the problem row and its own annotation (cascades).",
                "parameters": [
                    {
                        "description": "Problem ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.SuccessResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete a problem",
                "tags": [
                    "problems"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "Problem ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_problems.ProblemDetail"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Get a problem",
                "tags": [
                    "problems"
                ]
            },
            "put": {
                "description": "Allowed for admins (Council/Associate title) on any problem, or the problem's own creator. boulder_id is not editable here -- a boulder merge is how a problem changes boulders.",
                "parameters": [
                    {
                        "description": "Problem ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_problems.UpdateProblemRequest",
                                        "summary": "body",
                                        "description": "Updated fields"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Updated fields",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_problems.ProblemRow"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "not the creator and not an admin"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update a problem",
                "tags": [
                    "problems"
                ]
            }
        },
        "/api/problems/{id}/annotations": {
            "get": {
                "parameters": [
                    {
                        "description": "Problem ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/internal_problems.AnnotationRecord"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List a problem's topo annotations",
                "tags": [
                    "problems"
                ]
            },
            "put": {
                "description": "One vector-shape overlay per (problem, image_url). Allowed for admins or the problem's own creator.",
                "parameters": [
                    {
                        "description": "Problem ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_problems.SaveAnnotationRequest",
                                        "summary": "body",
                                        "description": "Image URL and shape data"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Image URL and shape data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_problems.AnnotationRecord"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "problem or image not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Save a topo image's annotation",
                "tags": [
                    "problems"
                ]
            }
        },
...
        "/api/upload/avatar": {
            "post": {
                "description": "Streams an image to Cloudinary's kepalabatu_avatars folder, returns its URL.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "avatar",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "Image file, max 8MB",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_problems.AvatarUploadResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "no file, file too large, or not an image"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "cloudinary upload failed"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Upload a profile avatar",
                "tags": [
                    "problems"
                ]
            }
        },
        "/api/upload/topo": {
            "post": {
                "description": "Streams an image to Cloudinary's kepalabatu_topos folder, returns its URL.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "image",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "Image file, max 8MB",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_problems.TopoUploadResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "no file, file too large, or not an image"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "cloudinary upload failed"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Upload a topo photo",
                "tags": [
                    "problems"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:3001/"
        }
    ]
}